requirements.txt
flask==2.2.2
pyodbc==4.0.39
SQLAlchemy==2.0.21
werkzeug===2.2.3 

azure-identity==1.7.1
azure-keyvault-secrets==4.3.0




